fix(centrifugo): fix superfluous error on response writer
- Add new control flow to Auth handler to fix the superfluous
response writer error/bug

[Finishes]
