+++ realtime/test_rtc.html
@@ -7,7 +7,7 @@
         <script src="https://cdn.jsdelivr.net/gh/centrifugal/centrifuge-js@2.8.0/dist/centrifuge.min.js"></script>
         <script type="text/javascript">
             const container = document.getElementById('counter')
-            const centrifuge = new Centrifuge("wss://realtime.zuri.chat/connection/websocket");
+            const centrifuge = new Centrifuge("ws://localhost:8000/connection/websocket");
             
             centrifuge.on('connect', function(ctx) {
                 console.log("connected", ctx);
